TG-30 Rewrite backend using database
Implemented new backend for jinya fonts based on mongodb and redis as cache for the binary files. Since font files are rather small we store them directly in the database instead of storing them in object store or the file system. To increase the access speed a redis cache is used.